Add quench time evolution plotting function and refactor related imports
